Исправление замечаний 1.0

+++ lesson1/task1.rb
@@ -3,7 +3,7 @@ name = gets.chomp
 name.capitalize!
 puts "Укажи свой рост:"
 height = gets.chomp
-ideal_height = (height.to_i - 110) * 1.15
+ideal_height = (height.to_f - 110) * 1.15
 
 if ideal_height >= 0
 	puts "#{name}, ваш идеальный вес #{ideal_height.round(2)} кг."
